+++ src/routes/shop/Shop.components.tsx
@@ -1,7 +1,7 @@
 import { Route, Routes } from 'react-router-dom'
 
 import ProductPreview from '../../components/product/product-preview/ProductPreview.component'
-import CategoriesPreview from '../categories-preview/CategoriesPreview.components'
+import CategoriesPreview from '../../components/category/categories-preview/CategoriesPreview.components'
 import Category from '../category/Category.component'
